partitioning: Add a new planner system
This new planner is undo-stack based designed with eventual UIs
in mind, and quite simple right now. Basically we track partitions/regions
and assert alignment boundaries, and test a bunch of common scenarios using
the new MockDisk APIs.

It will need some iteration as we go to account for usable LBA ranges within
the GPT headers, so we'll need to wrap the Planner up into a Strategy system
in a subsequent commit. Effectively we'll have the table in memory either as
a copy or a fresh one for a disk wipe.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>