partitioning: Add a new Strategy system
While the builtin strategies are arguably quite simple it does offer
quite enough reusable scaffolding that one can implement sane automatic
options for installers and provisioning.

The main focus is on disk wipes, and partition reuse. Such that if we
encounter an existing install we can amend to make it correct, or even
by manual planning APIs offer to delete a large rootfs and split into the
correct layout (xbootldr, anyone?)

In time we'll also support resizing operations but we'll need to get these
bits hooked up for GPT and partition types yet.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>