provisioning: Start the new KDL driven provisioning crate
This crate is mostly about loading a command queue from a KDL
configuration in order to dynamically produce strategies that
can be verified against disk planning APIs in the partitioning
crate(s).

For now our main focus is on building the command pieces, which
we've opted to bake in using `phf`, and leveraged `miette` to
ensure a Serpent OS project actually has decent error handling
from the start, for once. :P

Signed-off-by: Ikey Doherty <ikey@serpentos.com>